Allow canceling recurring reminders after execution for non-internal actors (#7337)
* Allow canceling recurring reminders after execution for non-internal actors

Internal actors have the ability to cancel a recurring reminder by responding to an execution with the special error `ErrReminderCanceled` after a reminder is executed.

However, the capability was never exposed to "external" actors.

This PR adds the ability for an actor to stop recurring reminders (and timers) after they are executed also for "external" actors. This is possible by including the header `X-DaprReminderCancel` with a truthy value in the response when a reminder or timer is executed.

This will require support in SDKs, but it allows something that is particularly helpful (and today can be buggy, see #6666)

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Lint

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

* Lint

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>

---------

Signed-off-by: ItalyPaleAle <43508+ItalyPaleAle@users.noreply.github.com>
Co-authored-by: Mukundan Sundararajan <65565396+mukundansundar@users.noreply.github.com>
Co-authored-by: Dapr Bot <56698301+dapr-bot@users.noreply.github.com>
Signed-off-by: Elena Kolevska <elena@kolevska.com>